fix(dashboard): visual-audit findings — calendar empty-state clip + Setup Guide pref + SQL cast
Caught while running the dev server locally and walking the dashboard:

Calendar — empty-state clipping rows:
- When filteredEvents.length===0, an EmptyState component rendered inside
  the calendar card and took flex-1, pushing the month grid up and
  clipping rows 3-6. In the screenshot Allan sent, only weeks 1-2 were
  visible; weeks 3-5 were hidden behind 'No events yet' overlay.
- Fix: only render the empty-state in WEEK view. In MONTH view the grid
  IS the canvas (click any day to create) — an empty-state overlay would
  always conflict with the grid layout.

Dashboard — Setup Guide collapse persistence:
- Was using safeGetItem/safeSetItem (localStorage). Allan reported it
  reopens on refresh — likely because his browser/incognito clears
  localStorage on session end.
- Moved to user_preferences.setup_guide_collapsed (boolean, defaults
  false). Now syncs across devices + survives any browser-cache reset.
- Added setup_guide_collapsed to PreferencesContext interface + defaults.
- Cleaned up unused safeGetItem/safeSetItem imports in Dashboard.tsx.

Health check SQL — Postgres ROUND signature:
- ROUND(double precision, integer) doesn't exist in Postgres — only
  ROUND(numeric, integer) does. AVG(EXTRACT(EPOCH FROM …)) returns
  double precision so was failing on first run.
- Added ::numeric casts on the duration calcs (avg_seconds + p95_seconds).

Verified via local dashboard dev server walk-through:
- Calendar month view: all 5 weeks render, day numbers full-visible
- Finances period filter: change to All Time → nav away → return → still All Time
- Tasks view toggle: switch to List → nav away → return → still List
- Setup Guide: click collapse → nav away → return → still collapsed
- Goals + Weekly Review seed data renders correctly

+++ dashboard/bathresurf-nsw-dashboard/supabase/health_check.sql
@@ -23,9 +23,10 @@ SELECT
     THEN ROUND(100.0 * COUNT(*) FILTER (WHERE status IN ('error','timeout')) / COUNT(*), 1)
     ELSE 0 END AS error_rate_pct,
   ROUND(COALESCE(SUM(cost_usd), 0), 2) AS total_cost_usd,
-  ROUND(COALESCE(AVG(EXTRACT(EPOCH FROM (completed_at - started_at))), 0), 1) AS avg_seconds,
+  -- Cast EXTRACT(EPOCH FROM …) to numeric — Postgres has no ROUND(double precision, int) overload.
+  ROUND(COALESCE(AVG(EXTRACT(EPOCH FROM (completed_at - started_at)))::numeric, 0), 1) AS avg_seconds,
   ROUND(COALESCE(
-    PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (completed_at - started_at))),
+    (PERCENTILE_CONT(0.95) WITHIN GROUP (ORDER BY EXTRACT(EPOCH FROM (completed_at - started_at))))::numeric,
     0), 1) AS p95_seconds
 FROM agent_runs
 WHERE started_at > now() - interval '24 hours';
